+++ CAENLabStatus/DZCLabTableViewManagerWithSublabs.m
@@ -70,7 +70,7 @@
         float usedPercent = [used floatValue] / [total floatValue];
         float freePercent = 1.0f - usedPercent;
 
-        cell.detailTextLabel.text = [NSString stringWithFormat:@"%d (%d%%) free", freeCount, (int)roundf(freePercent*100)];
+        cell.detailTextLabel.text = [NSString stringWithFormat:@"%d (%d%%) free", (int)freeCount, (int)roundf(freePercent*100)];
 
         if (usedPercent >= 0.8) {
             cell.textLabel.font = [UIFont systemFontOfSize:cell.textLabel.font.pointSize];
